+++ admin-ui/start.sh
@@ -79,6 +79,21 @@ http {
       proxy_set_header Connection "upgrade";
     }
 
+    location = /user {
+      return 302 /user/;
+    }
+
+    location ^~ /user/ {
+      proxy_pass http://127.0.0.1:8080;
+      proxy_http_version 1.1;
+      proxy_set_header Host \$host;
+      proxy_set_header X-Real-IP \$remote_addr;
+      proxy_set_header X-Forwarded-For \$proxy_add_x_forwarded_for;
+      proxy_set_header X-Forwarded-Proto \$scheme;
+      proxy_set_header Upgrade \$http_upgrade;
+      proxy_set_header Connection "upgrade";
+    }
+
     include ${PROXY_FILE};
   }
 }